chore: move Storybook stories in the projects\[id]\_components folder
