+++ src/app/(main)/projects/[id]/_components/contract-details.stories.tsx
@@ -1,12 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import ProjectInfoButton from "../projects/[id]/_components/project-info-button";
+import ProjectContractDetails from "./contract-details";
 import fakeProjectsData, {
     fakeEmptyContractDetails,
     fakeLongContractDetails,
 } from "./fake-data";
 const meta = {
-    title: "Main/Components/ProjectDetails/ProjectInfo",
-    component: ProjectInfoButton,
+    title: "Main/Components/ProjectDetails/Contract Details",
+    component: ProjectContractDetails,
     parameters: {
         layout: "fullscreen",
     },
@@ -18,26 +18,29 @@ const meta = {
         ),
     ],
     tags: ["autodocs"],
-} satisfies Meta<typeof ProjectInfoButton>;
+} satisfies Meta<typeof ProjectContractDetails>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const WithProjectDetails: Story = {
+export const WithContractDetails: Story = {
     args: {
+        id: fakeProjectsData.contractId,
         project: fakeProjectsData,
     },
 };
 
-export const WithLongProjectDetails: Story = {
+export const WithLongContractDetails: Story = {
     args: {
+        id: fakeLongContractDetails.contractId,
         project: fakeLongContractDetails,
     },
 };
 
-export const WithEmptyProjectDetails: Story = {
+export const WithoutContractDetails: Story = {
     args: {
+        id: fakeEmptyContractDetails.contractId,
         project: fakeEmptyContractDetails,
     },
 };
